I gota a fully featured deployment server up and.. running. You can now deploy packages to QA and once it passes, it will automatically deploy to production. The rollback feature is also implemented if it does not pass during the deployment process it will initiate a rollback to the previous version automatically as well.
I've also added the deploy script that is used inside dev or even QA to automatically bundle and deploy the package to the server. It needs to be configured with the correct values in the .ini file.
To call you must run the script with the following sample command: ./deploy /path/to/your/package/directory.